完成分页查询用户列表

+++ src/main/java/com/wanbang/resp/UserListResp.java
@@ -8,8 +8,8 @@ import java.util.List;
 //返回所有用户列表
 @Data
 public class UserListResp {
-    private int total;
-    private int page;
-    private int size;
+    private Long total;
+    private Long page;
+    private Long size;
     private List<UserInfoVO> items;
 }
